requirements.txt
bidict==0.22.1
blinker==1.6.3
click==8.1.7
Flask==3.0.0
Flask-Login==0.6.2
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.0
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
passlib==1.7.4
psycopg2==2.9.9
python-dotenv==1.0.0
python-engineio==4.8.0
python-socketio==5.10.0
simple-websocket==1.0.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
Werkzeug==3.0.1
wsproto==1.2.0
WTForms==3.1.0
